+++ Domain/Repositories/Articles/IArticleRepository.cs
@@ -4,9 +4,9 @@ namespace Domain.Repositories.Articles
 {
     public interface IArticleRepository
     {
-        Task<Article?> GetById(long id);
+        Task<Article?> GetById(int id);
         Task Insert(Article article);
         Task Update(Article article);
-        Task Delete(long id);
+        Task Delete(int id);
     }
 }
